requirements.txt
metis-agent>=0.1.3
flask>=2.0.0
python-dotenv>=0.15.0
requests>=2.25.0